fix compiler warning (from trac https://trac.osgeo.org/grass/ticket/714, comment 7)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/develbranch_6@40643 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/gis/spawn.c
@@ -705,7 +705,7 @@ static void begin_spawn(struct spawn *sp)
     sp->directory = NULL;
 }
 
-#define NEXT_ARG(var, type) ((type) *(var)++)
+#define NEXT_ARG(var, type) ((type) (ssize_t) *(var)++)
 
 static void parse_argvec(struct spawn *sp, const char **va)
 {
